Move BLM event fab to declaration file
