+++ src/parser/jobs/blm/modules/RotationWatchdog.tsx
@@ -7,7 +7,7 @@ import {ActionLink} from 'components/ui/DbLink'
 import {RotationTable} from 'components/ui/RotationTable'
 import {getDataBy} from 'data'
 import ACTIONS from 'data/ACTIONS'
-import {CastEvent, BaseEventFields} from 'fflogs'
+import {CastEvent} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import Checklist from 'parser/core/modules/Checklist'
 import Combatants from 'parser/core/modules/Combatants'
@@ -20,7 +20,7 @@ import UnableToAct from 'parser/core/modules/UnableToAct'
 import {EntityStatuses} from 'parser/core/modules/EntityStatuses'
 import DISPLAY_ORDER from './DISPLAY_ORDER'
 import {FIRE_SPELLS} from './Elements'
-import {BLM_GAUGE_EVENT} from './Gauge'
+import {BLM_GAUGE_EVENT, BLMGaugeEvent} from './Gauge'
 
 const DEBUG_SHOW_ALL_CYCLES = false && process.env.NODE_ENV !== 'production'
 
@@ -142,26 +142,6 @@ class Cycle {
 
 	public overrideErrorCode(code: {priority: number, message: TODO}) {
 		this._errorCode = code
-	}
-}
-
-// TS typedef for BLM Gauge events so it doesn't choke
-// TODO: Move to Gauge if that gets converted to TS
-interface BLMGaugeEvent extends BaseEventFields {
-	type: typeof BLM_GAUGE_EVENT,
-	timestamp: number,
-	insertAfter: number,
-	astralFire: number,
-	umbralIce: number,
-	umbralHearts: number,
-	enochian: boolean,
-	polyglot: number,
-	lastGaugeEvent: BLMGaugeEvent,
-}
-
-declare module 'events' {
-	interface EventTypeRepository {
-		blmRotationWatchdog: BLMGaugeEvent
 	}
 }
